+++ packages/graph/src/find-path.ts
@@ -7,22 +7,28 @@ export const findShortestPath = (
   start: string,
   end: string,
   directed?: boolean,
-  weightPropertyName?: string
+  weightPropertyName?: string,
 ) => {
-  const { length, path, allPath } = dijkstra(
+  const { length, path, allPath, edgePath, allEdgePath } = dijkstra(
     graphData,
     start,
     directed,
-    weightPropertyName
+    weightPropertyName,
   );
-  return { length: length[end], path: path[end], allPath: allPath[end] };
+  return {
+    length: length[end],
+    path: path[end],
+    allPath: allPath[end],
+    edgePath: edgePath[end],
+    allEdgePath: allEdgePath[end],
+  };
 };
 
 export const findAllPath = (
   graphData: GraphData,
   start: string,
   end: string,
-  directed?: boolean
+  directed?: boolean,
 ) => {
   if (start === end) return [[start]];
 
@@ -32,9 +38,7 @@ export const findAllPath = (
   const isVisited = { [start]: true };
   const stack: string[][] = []; // 辅助栈，用于存储访问过的节点的邻居节点
   const allPath = [];
-  let neighbors = directed
-    ? getNeighbors(start, edges, 'target')
-    : getNeighbors(start, edges);
+  let neighbors = directed ? getNeighbors(start, edges, 'target') : getNeighbors(start, edges);
   stack.push(neighbors);
 
   while (visited.length > 0 && stack.length > 0) {
@@ -44,10 +48,8 @@ export const findAllPath = (
       if (child) {
         visited.push(child);
         isVisited[child] = true;
-        neighbors = directed
-          ? getNeighbors(child, edges, 'target')
-          : getNeighbors(child, edges);
-        stack.push(neighbors.filter(neighbor => !isVisited[neighbor]));
+        neighbors = directed ? getNeighbors(child, edges, 'target') : getNeighbors(child, edges);
+        stack.push(neighbors.filter((neighbor) => !isVisited[neighbor]));
       }
     } else {
       const node = visited.pop();
@@ -57,7 +59,7 @@ export const findAllPath = (
     }
 
     if (visited[visited.length - 1] === end) {
-      const path = visited.map(node => node);
+      const path = visited.map((node) => node);
       allPath.push(path);
 
       const node = visited.pop();